third_party/blink/web_tests/external/wpt/css/css-inline/text-box-trim/parsing/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-box-edge inherits
  assert_equals: expected "cap" but got "auto"
Harness: the test ran to completion.
